Introduce "//guide" internal links
These link to another guide within the current org and version.

While here, drop the need for these internal links to begin with a
leading underscore. The link types without the underscore look cleaner,
and given that these URLs appear as scheme-less with the link types as
their host, there's no need to disambiguate them further.