Fix bug where toc header text would go missing
The problem was to do with Markdown-formatted headers that may contain multiple text chunks.

For example, a header of “Hello (`world`)” would yield text chunks of “Hello (“, “world”, and “)”, because that middle “world” chunk lived inside a nested code element.

Previously, our linkable headings filter was a little naive: it expexted a single text chunk per header and would directly assign each yielded text chunk to the header, resulting in the final chunk “winning” and becoming the text for the header.

Now, we concat each chunk as it comes, which results in properly complete header text for headers like this.